type checking issues patched
