+++ pyproject.toml
@@ -34,18 +34,17 @@ classifiers = [
 # DEPENDENCIES
 
 [tool.poetry.dependencies]
-python = ">=3.8.1,<3.13"
-hyperopt = "0.2.7"
-numpy = "1.24.4"
-packaging = "23.1"
-pandas = "2.0.1"
+python = ">=3.9,<3.13"
+hyperopt = "*"
+numpy = ">= 1.24"
+pandas = ">= 2.0.1"
 scipy = "*"
-scikit-learn = "1.3.2"
+scikit-learn = ">= 1.6"
 sphinx-markdown-tables = { version = "*", optional = true }
-statsmodels = "0.14.0"
+statsmodels = ">= 0.14.0"
 typed-ast = { version = "*", optional = true }
 category-encoders = "^2.6.3"
-dcor = "0.6"
+dcor = ">= 0.6"
 
 [tool.poetry.group.torch.dependencies]
 torch = "< 2.5"
@@ -57,6 +56,7 @@ jupyter = "1.0.0"
 jupyterlab = "1.2.6"
 jupytext = "1.14.4"
 matplotlib = "3.6.2"
+packaging = "23.1"
 pre-commit = "2.21.0"
 twine = "3.7.1"
 wheel = "0.37.1"
