+++ qolmat/benchmark/missing_patterns.py
@@ -371,7 +371,8 @@ class _SamplerHoleGenerator(_HoleGenerator):
                 sample = min(min(sample, sizes_max.max()), n_masked_left)
                 i_hole = self.rng.choice(np.where(sample <= sizes_max)[0])
 
-                if not (~mask[column].iloc[i_hole - sample : i_hole]).all():
+                indices_hole = mask.index[i_hole - sample : i_hole]
+                if not (~mask.loc[indices_hole, column]).all():
                     raise ValueError(
                         "The mask condition is not satisfied for "
                         f"column={column}, "
@@ -379,7 +380,7 @@ class _SamplerHoleGenerator(_HoleGenerator):
                         f"and i_hole={i_hole}."
                     )
 
-                mask[column].iloc[i_hole - sample : i_hole] = True
+                mask.loc[indices_hole, column] = True
                 n_masked_left -= sample
 
                 sizes_max.iloc[i_hole - sample : i_hole] = 0
